+++ nisystemlink/clients/testmonitor/models/__init__.py
@@ -1,17 +1,33 @@
 from ._api_info import Operation, V2Operations, ApiInfo
+from ._named_value import NamedValue
 from ._result import Result
 from ._status import StatusType, Status
-from ._create_results_partial_success import CreateResultsPartialSuccess
-from ._update_results_partial_success import UpdateResultsPartialSuccess
-from ._delete_results_partial_success import DeleteResultsPartialSuccess
+from ._step import Step
+from ._step_data import Measurement, StepData
 from ._paged_results import PagedResults
+from ._paged_steps import PagedSteps
+from ._delete_results_partial_success import DeleteResultsPartialSuccess
+from ._delete_steps_partial_success import DeleteStepsPartialSuccess, StepIdResultIdPair
+from ._create_results_partial_success import CreateResultsPartialSuccess
+from ._create_steps_partial_success import CreateStepsPartialSuccess
+from ._update_results_partial_success import UpdateResultsPartialSuccess
+from ._update_steps_partial_success import UpdateStepsPartialSuccess
 from ._create_result_request import CreateResultRequest
+from ._create_steps_request import CreateStepRequest
 from ._update_result_request import UpdateResultRequest
+from ._update_steps_request import UpdateStepRequest
 from ._query_results_request import (
     QueryResultsRequest,
     QueryResultValuesRequest,
     ResultField,
     ResultProjection,
+)
+from ._query_steps_request import (
+    QueryStepsRequest,
+    QueryStepValuesRequest,
+    StepOrderBy,
+    StepField,
+    StepProjection,
 )
 
 # flake8: noqa