+++ nisystemlink/clients/testmonitor/models/_named_value.py
@@ -1,0 +1,11 @@
+from typing import Any
+
+from nisystemlink.clients.core._uplink._json_model import JsonModel
+
+
+class NamedValue(JsonModel):
+    name: str
+    """The name of the value."""
+
+    value: Any
+    """The value."""